--- Electron/src/componentes/Sidebar/Playlist/ContextMenuSong/contextMenuPlaylist.module.css
@@ -1,82 +1,0 @@
-.wrapperContextMenu {
-
-
-  background-color: var(--hover-white);
-  width: 250px;
-  border-radius: 4px;
-  color: hsla(0, 0%, 100%, .9);
-  padding: 4px
-}
-
-.wrapperContextMenu ul {
-
-  margin: 0px;
-  padding: 1% !important;
-
-}
-
-.wrapperContextMenu ul li {
-
-  padding: 1% !important;
-  border-top: 1px solid var(--separator-white);
-  border-bottom: 1px solid var(--separator-white);
-
-}
-
-
-.wrapperContextMenu ul li:first-child {
-
-  border: none;
-  border-bottom: 1px solid var(--separator-white);
-
-}
-
-.wrapperContextMenu ul li:last-child {
-
-  border: none !important;
-}
-
-.wrapperContextMenu ul li+li {
-  border-top: none;
-  /* Elimina el borde superior del siguiente <li> */
-}
-
-
-.wrapperContextMenu ul li button {
-
-  background-color: transparent;
-  border: none;
-  width: 100%;
-  display: flex;
-  justify-content: flex-start;
-  color: inherit;
-  font-size: 0.875rem;
-  padding: 12px;
-  border-radius: 2px;
-  text-align: start;
-  /* font-weight: 300 !important;
- */
-
-}
-
-.wrapperContextMenu ul button:hover {
-
-  background-color: hsla(0, 0%, 100%, .1);
-
-}
-
-
-/* Add to playlist */
-
-.wrapperContextMenuAddToPlaylist{
-
-  max-width: 350px !important;
-  max-height: 40svh;
-}
-
-.wrapperContextMenuAddToPlaylist li:not(:nth-child(2)){
-
-  border: none !important;
-
-}
-
